+++ packages/axway-cli-oum/src/lib/util.js
@@ -1,6 +1,19 @@
 import { initSDK } from '@axway/amplify-cli-utils';
 
 /**
+ * Formats a date in the format "m/d/yyyy".
+ *
+ * @param {Date|Number} dt - The date to format.
+ * @returns {String}
+ */
+export function formatDate(dt) {
+	if (!(dt instanceof Date)) {
+		dt = new Date(dt);
+	}
+	return `${dt.getUTCMonth() + 1}/${dt.getUTCDate()}/${dt.getUTCFullYear()}`;
+}
+
+/**
  * Initializes the Amplify SDK, loads an account, and finds the default org id.
  *
  * @param {String} [accountName] - The name of the platform account to use.